Remove additional questions core component
