+++ packages/webcomponents/src/components/business-forms/payment-provisioning/readme.md
@@ -57,11 +57,10 @@ graph TD;
   justifi-legal-address-form-step-core --> justifi-form-address-fields
   justifi-form-address-fields --> form-control-text
   justifi-form-address-fields --> form-control-select
-  justifi-additional-questions-form-step --> justifi-additional-questions-form-step-core
-  justifi-additional-questions-form-step-core --> form-control-tooltip
-  justifi-additional-questions-form-step-core --> form-control-monetary-provisioning
-  justifi-additional-questions-form-step-core --> form-control-select
-  justifi-additional-questions-form-step-core --> form-control-text
+  justifi-additional-questions-form-step --> form-control-tooltip
+  justifi-additional-questions-form-step --> form-control-monetary-provisioning
+  justifi-additional-questions-form-step --> form-control-select
+  justifi-additional-questions-form-step --> form-control-text
   justifi-business-representative-form-step --> justifi-business-representative-form-step-core
   justifi-business-representative-form-step-core --> justifi-business-representative-form-inputs
   justifi-business-representative-form-inputs --> form-control-tooltip
